Start cleaning up, remove all traces of old emb_display
