+++ source/melpomene/src/sim_drivers.rs
@@ -1,3 +1,2 @@
-// pub mod emb_display;
-pub mod emb_display2;
+pub mod emb_display;
 pub mod tcp_serial;
